Audio/ValKit: Moved the AudioTestWaveXXX functions to AudioTest.cpp so those can also be used for injecting input (recording) data later [build fixes]. bugref:10008
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@89073 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/AudioTest.cpp
@@ -1070,14 +1070,14 @@ int AudioTestWaveFileRead(PAUDIOTESTWAVEFILE pWaveFile, void *pvBuf, size_t cbBu
     {
         if (pcbRead)
         {
-            pWaveFile->offCur += *pcbRead;
+            pWaveFile->offCur += (uint32_t)*pcbRead;
             if (cbBuf > *pcbRead)
                 rc = VINF_EOF;
             else if (!cbBuf && pWaveFile->offCur == pWaveFile->cbSamples)
                 rc = VINF_EOF;
         }
         else
-            pWaveFile->offCur += cbBuf;
+            pWaveFile->offCur += (uint32_t)cbBuf;
     }
     return rc;
 }
